Move tests into folder
