+++ shopsafe-frontend/src/app/classes/store/store-data.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
-import { ApiService } from '../../../api/api.service';
-import { Observable, throwError } from 'rxjs';
+import { ApiService } from '../../api/api.service';
+import { Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
