+++ shopsafe-frontend/src/app/tests/about.component.spec.ts
