+++ shopsafe-frontend/src/app/tests/api.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ApiService } from './api.service';
+import { ApiService } from '../api/api.service';
 
 describe('ApiService', () => {
   let service: ApiService;
