+++ shopsafe-frontend/src/app/tests/app.component.spec.ts
