+++ shopsafe-frontend/src/app/tests/check-in-modal.component.spec.ts
