+++ shopsafe-frontend/src/app/tests/landing.component.spec.ts
