+++ shopsafe-frontend/src/app/tests/result.component.spec.ts
