+++ shopsafe-frontend/src/app/tests/result.spec.ts
