+++ shopsafe-frontend/src/app/tests/store-data.service.spec.ts
