+++ shopsafe-frontend/src/app/tests/store.component.spec.ts
