+++ shopsafe-frontend/src/app/tests/store.spec.ts
